big increment

+++ database/migrations/2014_10_11_054519_create_departaments_table.php
@@ -14,10 +14,10 @@ class CreateDepartamentsTable extends Migration
     public function up()
     {
         Schema::create('departaments', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('nombre');
             $table->boolean('activo')->default(1);
-            $table->integer('id_lider');
+            $table->bigInteger('id_lider');
             $table->timestamps();
         });
     }
