+++ database/migrations/2014_10_11_055536_create_equipment_table.php
@@ -14,9 +14,9 @@ class CreateEquipmentTable extends Migration
     public function up()
     {
         Schema::create('equipment', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('nombre');
-            $table->integer('departaments_id')->unsigned()->nullable($value = true);
+            $table->bigInteger('departaments_id')->unsigned()->nullable($value = true);
             $table->foreign('departaments_id')->references('id')->on('departaments')->onDelete('cascade');
             $table->timestamps();
         });
